requirements.txt
ollama==0.3.3
playwright==1.48.0
Pillow==10.4.0